Remove Unused Script
Removed the original test version of JSONReader.cs, and renamed JSONReader2.cs. Also deactivated Item.cs changes for pull to master.
