+++ MARDEK Engine/Assets/ScriptableObjects/Item/PlotItems/Dark Crystal of Belfan.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: Dark Crystal of Belfan
   m_EditorClassIdentifier: 
   _displayName: Dark Crystal of Belfan
-  _elementText: Dark
+  _elementText: 
   _description: An Artefact of the Gods. It is essentially crystalised elemental
     energy, and one of the planet's 'organs', of a sort. This particular crystal
     is the source of all malice and greed on Belfan.
