+++ MARDEK Engine/Assets/Scripts/MARDEK/Inventory/Item/Item.cs
@@ -10,17 +10,19 @@ namespace MARDEK.Inventory
     public class Item : AddressableScriptableObject
     {
         [SerializeField] string _displayName;
-        public string displayName { get { return _displayName; } set { _displayName = value;} }
-        [SerializeField] string _elementText;
-        public string elementText { get { return _elementText; } set { _elementText = value;} }
+        public string displayName { get { return _displayName; } }
+        // Added only to fascilitate mass inputs. Should not be active unless you are attempting to create items via JSONReader.cs
+        //[SerializeField] string _elementText;
+        //public string elementText { get { return _elementText; } }
         [SerializeField] string _description;
-        public string description { get { return CreateFullDescription(_description); } set { _description = value;} }
+        public string description { get { return CreateFullDescription(_description); } }
         [SerializeField] Sprite _sprite;
         public Sprite sprite { get { return _sprite; } }
         [SerializeField] Element _element;
-        public Element element { get { return _element; } set { _element = value;}  }
+        // Setter is needed to mass import items, but should be unavailable for normal development
+        public Element element { get { return _element; } /*set{ _element = value;}*/ }
         [SerializeField] int _price;
-        public int price { get { return _price; }  set { _price = value;} }
+        public int price { get { return _price; } }
         public string properties { get { return CreateProperties(); } }
 
         [field: SerializeField] public List<Skill.Skill> SkillsToEquip { get; private set; }
