+++ MARDEK Engine/Assets/Scripts/MARDEK/Inventory/Item/JSONReader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4bd5342d3b8047a4abafff3a5efd2d5a
+guid: d0d4e7c93356d7c45b8fff0a4f53dd00
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
