--- MARDEK Engine/Assets/Scripts/MARDEK/Inventory/Item/JSONReader2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d0d4e7c93356d7c45b8fff0a4f53dd00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
